internetRadio-1.3.0-wifi-manager/src/network/wifi_manager.py
import subprocess
import logging
import time
from typing import List, Dict, Optional, Any
import os
from src.utils.config_manager import ConfigManager
import socket
import shutil

class WiFiManager:
    def __init__(self):
        self.logger = logging.getLogger('network')
        self.initialized = False
        
    def initialize(self) -> bool:
        """Initialize WiFi manager"""
        try:
            self.initialized = True
            return True
        except Exception as e:
            self.logger.error(f"Error initializing WiFi manager: {e}")
            return False
            
    def connect_to_network(self, ssid: str, password: str) -> bool:
        """Connect to a WiFi network"""
        try:
            return True
        except Exception as e:
            self.logger.error(f"Error connecting to network: {e}")
            return False
            
    def is_connected(self) -> bool:
        """Check if connected to a network"""
        return True
        
    def cleanup(self) -> None:
        """Cleanup WiFi resources"""
        pass